major edit to progam and additional scripts
-main edits involve increasing efficiency and minor timing changes
- additional script for generating new target files with full target space (new tgt file demo_r4 included)

+++ TargetFile.py
@@ -22,7 +22,7 @@ class TargetFile:
                 s = "currTrialStr"
                 setattr(self,s,row)
                 for value,index in zip(row,self.headers):
-                    setattr(self,index,int(value))
+                    setattr(self,index,float(value))
 
     # method to increment the trial number
     def nextTrial(self):
